+++ src/main/java/com/crmdemo/entity/CrmpraiseComplaint.java
@@ -3,49 +3,49 @@ package com.crmdemo.entity;
 
 public class CrmpraiseComplaint {
 
-  private long id;
-  private long userId;
-  private long complaintStatus;
-  private long complaintReason;
+  private Integer id;
+  private Integer userId;
+  private Integer complaintStatus;
+  private Integer complaintReason;
   private String otherReason;
-  private long disposeStatus;
+  private Integer disposeStatus;
   private java.sql.Timestamp disposeTime;
   private java.sql.Timestamp updateTime;
   private java.sql.Timestamp createTime;
 
 
-  public long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
 
-  public long getUserId() {
+  public Integer getUserId() {
     return userId;
   }
 
-  public void setUserId(long userId) {
+  public void setUserId(Integer userId) {
     this.userId = userId;
   }
 
 
-  public long getComplaintStatus() {
+  public Integer getComplaintStatus() {
     return complaintStatus;
   }
 
-  public void setComplaintStatus(long complaintStatus) {
+  public void setComplaintStatus(Integer complaintStatus) {
     this.complaintStatus = complaintStatus;
   }
 
 
-  public long getComplaintReason() {
+  public Integer getComplaintReason() {
     return complaintReason;
   }
 
-  public void setComplaintReason(long complaintReason) {
+  public void setComplaintReason(Integer complaintReason) {
     this.complaintReason = complaintReason;
   }
 
@@ -59,11 +59,11 @@ public class CrmpraiseComplaint {
   }
 
 
-  public long getDisposeStatus() {
+  public Integer getDisposeStatus() {
     return disposeStatus;
   }
 
-  public void setDisposeStatus(long disposeStatus) {
+  public void setDisposeStatus(Integer disposeStatus) {
     this.disposeStatus = disposeStatus;
   }
